# perf: improve scheduling speed