Allocatable() is called a lot. Memoize this using a sync.Map so we can avoid the expensive calculation.

name               old time/op   new time/op   delta
Scheduling1-12       566µs ± 9%    382µs ± 5%  -32.54%  (p=0.004 n=5+6)
Scheduling50-12     26.2ms ± 4%   20.0ms ± 6%  -23.53%  (p=0.004 n=5+6)
Scheduling100-12    53.1ms ± 8%   39.4ms ± 6%  -25.70%  (p=0.004 n=5+6)
Scheduling500-12     272ms ± 5%    202ms ±10%  -25.68%  (p=0.004 n=5+6)
Scheduling1000-12    596ms ±11%    457ms ±12%  -23.38%  (p=0.004 n=5+6)
Scheduling2000-12    1.08s ± 4%    0.89s ±10%  -17.46%  (p=0.004 n=5+6)
Scheduling5000-12    2.83s ± 8%    2.21s ± 6%  -21.91%  (p=0.004 n=5+6)

<!-- Please follow the guidelines at https://www.conventionalcommits.org/en/v1.0.0/ and use one of the following in your title:
feat:            <-- New features that require a MINOR version update
fix:             <-- Bug fixes that require at PATCH version update
chore:           <-- Smaller changes that impact behavior but aren't large enough to be features
perf:            <-- Code changes that improve performance but do not impact behavior
docs:            <-- Documentation changes that do not impact code
test:            <-- Test changes that do not impact behavior
ci:              <-- Changes that affect test or rollout automation
!${type}:        <-- Include ! if your change includes a backwards incompatible change.
-->

Fixes #N/A <!-- issue number -->

**Description**

**How was this change tested?**

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.